requirements.txt
# Core framework
fastapi==0.110.0
uvicorn==0.29.0

# Database & ORM
sqlalchemy==2.0.29
psycopg2-binary==2.9.9        # If using PostgreSQL
# sqlite needs no extra package

# Auth & Security
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# Environment variables
python-dotenv==1.0.1

# CORS & Requests
requests==2.31.0

# OpenAI (AI Recommendations)
openai==1.14.3

# Pydantic (FastAPI models)
pydantic==2.6.4
pydantic-settings==2.2.1

# Utilities
email-validator==2.1.0.post1
